# ml-close-loop-be/openapi.yaml
openapi: 3.1.0
info:
  title: DEFNEX MLOps Backend
  version: 0.1.0
paths:
...
  /api/v1/audit-logs:
    get:
      tags:
      - Audit
      summary: List Audit Logs
      description: 'Read-only audit trail (issue #129), admin only. Filterable by
        actor, action, resource

        (type and/or id), and a created_at date range; reuses `get_pagination` for
        page/size.'
      operationId: list_audit_logs_api_v1_audit_logs_get
      parameters:
      - name: actor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Actor Id
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - name: resource_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Type
      - name: resource_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_AuditLogEntry_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: List Users
      operationId: list_users_api_v1_users_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_UserResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/datasets:
    get:
      tags:
      - Datasets
      summary: List Datasets
      operationId: list_datasets_api_v1_datasets_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_DatasetSummary_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/eval-sets:
    get:
      tags:
      - Eval Sets
      summary: List Eval Sets
      operationId: list_eval_sets_api_v1_eval_sets_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvalSetSummary'
                title: Response List Eval Sets Api V1 Eval Sets Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/feedback:
    post:
      tags:
      - Feedback
      summary: Submit Feedback
      description: 'Submit feedback on an inference response (issue #42). `model_id`/`version`
        must be an

        existing model version - `get_model_version_or_404` (issue #36''s helper)
        resolves it the

        same way every other model-version-scoped endpoint does.'
      operationId: submit_feedback_api_v1_feedback_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackRecord'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Feedback
      summary: List Feedback
      description: '`fl.status` filters on curation_status (PENDING/APPROVED/REJECTED);
        `fl.model` filters

        on the referenced model_id - same `FilterParams`/`PaginationParams` every
        other list

        endpoint uses, no new pagination/filter logic (issue #42''s own explicit requirement).'
      operationId: list_feedback_api_v1_feedback_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_FeedbackRecord_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/training-runs:
    post:
      tags:
      - Training
      summary: Create Training Run
      operationId: create_training_run_api_v1_training_runs_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingRunCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingRun'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Training
      summary: List Training Runs
      operationId: list_training_runs_api_v1_training_runs_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_TrainingRun_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/transfers:
    post:
      tags:
      - Transfers
      summary: Initiate Transfer
      description: Initiate an artifact transfer (PRD §20.3). Admin only.
      operationId: initiate_transfer_api_v1_transfers_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferInitiateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Transfers
      summary: List Transfers
      description: List transfers. Any authenticated user.
      operationId: list_transfers_api_v1_transfers_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_Transfer_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/compute-resources:
    post:
      tags:
      - Compute Resources
      summary: Create Compute Resource
      description: Register a new compute resource (PRD §19.4). Admin only.
      operationId: create_compute_resource_api_v1_compute_resources_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComputeResourceCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputeResource'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Compute Resources
      summary: List Compute Resources
      description: List compute resources with optional filters. Any authenticated
        user.
      operationId: list_compute_resources_api_v1_compute_resources_get
      parameters:
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      - name: environment
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ComputeResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/models:
    get:
      tags:
      - Models
      summary: List Models
      operationId: list_models_api_v1_models_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: model
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
      - name: authorization
        in: header
        required: false
        schema:
          type: string
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelSummary'
                title: Response List Models Api V1 Models Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
      description: 'openapi.yaml Error — `{"error": {"code": ..., "message": ...}}`.'
...
    EvalSetSummary:
      properties:
        eval_set_id:
          type: string
          title: Eval Set Id
        latest_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Latest Version
        version_count:
          type: integer
          title: Version Count
          default: 0
      type: object
      required:
      - eval_set_id
      title: EvalSetSummary
      description: Lightweight list-view entry for GET /eval-sets (openapi.yaml EvalSetSummary).
...
    FeedbackCreateRequest:
      properties:
        model_id:
          type: string
          title: Model Id
        version:
          type: integer
          title: Version
        prompt:
          type: string
          minLength: 1
          title: Prompt
        response:
          type: string
          minLength: 1
          title: Response
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
        correction:
          anyOf:
          - type: string
          - type: 'null'
          title: Correction
      type: object
      required:
      - model_id
      - version
      - prompt
      - response
      - rating
      title: FeedbackCreateRequest
      description: 'Submit feedback on an inference response (issue #42).


        `model_id`/`version` identify the concrete `ModelVersion` that produced `response`

        (`InferenceResponse.model_id`/`version`, issue #41) - not an alias, so the
        reference is

        traceable to an exact version regardless of what is DEPLOYED later. `rating`
        is a plain

        1-5 input-validation bound (a common feedback scale), not an invented evaluation
        metric.

        `correction` is optional - a human-authored better answer, when the reviewer
        has one.'
...
    FeedbackRecord:
      properties:
        feedback_id:
          type: string
          title: Feedback Id
        model_id:
          type: string
          title: Model Id
        version:
          type: integer
          title: Version
        prompt:
          type: string
          title: Prompt
        response:
          type: string
          title: Response
        rating:
          type: integer
          title: Rating
        correction:
          anyOf:
          - type: string
          - type: 'null'
          title: Correction
        curation_status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
          title: Curation Status
        submitted_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Submitted By
        submitted_at:
          type: string
          format: date-time
          title: Submitted At
        curated_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Curated By
        curated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Curated At
      type: object
      required:
      - feedback_id
      - model_id
      - version
      - prompt
      - response
      - rating
      - curation_status
      - submitted_at
      title: FeedbackRecord
      description: A feedback row (openapi.yaml Feedback).
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_AuditLogEntry_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AuditLogEntry'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - pages
      title: PaginatedResponse[AuditLogEntry]
...
    PaginatedResponse_DatasetSummary_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DatasetSummary'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - pages
      title: PaginatedResponse[DatasetSummary]
    PaginatedResponse_FeedbackRecord_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/FeedbackRecord'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - pages
      title: PaginatedResponse[FeedbackRecord]
...
    PaginatedResponse_TrainingRun_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TrainingRun'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - pages
      title: PaginatedResponse[TrainingRun]
    PaginatedResponse_Transfer_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Transfer'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - pages
      title: PaginatedResponse[Transfer]
    PaginatedResponse_UserResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - pages
      title: PaginatedResponse[UserResponse]
...
    TrainingRunCreateRequest:
      properties:
        dataset_id:
          type: string
          title: Dataset Id
        dataset_version:
          type: integer
          title: Dataset Version
        model_id:
          type: string
          title: Model Id
        base_model:
          type: string
          title: Base Model
        training_config:
          $ref: '#/components/schemas/TrainingConfig'
        triggered_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Triggered By
        compute_resource_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Compute Resource Id
      type: object
      required:
      - dataset_id
      - dataset_version
      - model_id
      - base_model
      - training_config
      title: TrainingRunCreateRequest
      description: Request to create a training run (openapi.yaml TrainingRunCreateRequest).
...
    TransferInitiateRequest:
      properties:
        artifact_uri:
          type: string
          title: Artifact Uri
        source_host:
          type: string
          title: Source Host
        target_host:
          type: string
          title: Target Host
      type: object
      required:
      - artifact_uri
      - source_host
      - target_host
      title: TransferInitiateRequest
      description: Request to initiate an artifact transfer.
...